rename and clean up examples
