+++ source/PerformanceTests/ExamplePerformanceTests/NonDiscoverable/DisabledTestIsIgnoredExample.cs
@@ -4,7 +4,7 @@ using Sailfish.Attributes;
 namespace PerformanceTests.ExamplePerformanceTests.NonDiscoverable;
 
 [Sailfish(Disabled = true)]
-public class DisabledTestIsNotDiscoverable
+public class DisabledTestIsIgnoredExample
 {
     [SailfishMethod]
     public void MethodShouldNotHavePlayButton()
