--- source/PerformanceTests/ExceptionHandlingExamples/IterationSetupExceptionComesFirst.cs
@@ -1,37 +1,0 @@
-using System;
-using System.Threading;
-using System.Threading.Tasks;
-using Sailfish.Attributes;
-
-namespace PerformanceTests.ExceptionHandlingExamples;
-
-[Sailfish(NumIterations = 1, NumWarmupIterations = 1, Disabled = false)]
-public class IterationSetupExceptionComesFirst
-{
-    [SailfishMethod]
-    public async Task LifeCycleExceptionTests(CancellationToken cancellationToken)
-    {
-        await Task.Delay(10, cancellationToken);
-    }
-
-    [SailfishIterationSetup]
-    public async Task IterationSetup(CancellationToken cancellationToken)
-    {
-        await Task.Delay(10, cancellationToken);
-        throw new Exception("Iteration Setup Exception");
-    }
-
-    [SailfishMethod]
-    public async Task SailfishMethodException(CancellationToken cancellationToken)
-    {
-        await Task.Delay(10, cancellationToken);
-        throw new Exception("Sailfish Method Exception");
-    }
-
-    [SailfishMethodTeardown]
-    public async Task MethodTeardown(CancellationToken cancellationToken)
-    {
-        await Task.Delay(10, cancellationToken);
-        throw new Exception("Method Teardown Exception");
-    }
-}
