--- source/PerformanceTests/ExceptionHandlingExamples/IterationSetupExceptionIsHandled.cs
@@ -1,23 +1,0 @@
-using System;
-using System.Threading;
-using System.Threading.Tasks;
-using Sailfish.Attributes;
-
-namespace PerformanceTests.ExceptionHandlingExamples;
-
-[Sailfish(NumIterations = 1, NumWarmupIterations = 1, Disabled = false)]
-public class IterationSetupExceptionIsHandled
-{
-    [SailfishIterationSetup]
-    public async Task IterationSetup(CancellationToken cancellationToken)
-    {
-        await Task.Delay(10, cancellationToken);
-        throw new Exception("Iteration Setup Exception");
-    }
-
-    [SailfishMethod]
-    public async Task LifeCycleExceptionTests(CancellationToken cancellationToken)
-    {
-        await Task.Delay(10, cancellationToken);
-    }
-}
