--- source/PerformanceTests/ExceptionHandlingExamples/MethodTeardownExceptionIsHandled.cs
@@ -1,23 +1,0 @@
-using System;
-using System.Threading;
-using System.Threading.Tasks;
-using Sailfish.Attributes;
-
-namespace PerformanceTests.ExceptionHandlingExamples;
-
-[Sailfish(NumIterations = 1, NumWarmupIterations = 1, Disabled = false)]
-public class MethodTeardownExceptionIsHandled
-{
-    [SailfishMethod]
-    public async Task LifeCycleExceptionTests(CancellationToken cancellationToken)
-    {
-        await Task.Delay(10, cancellationToken);
-    }
-
-    [SailfishMethodTeardown]
-    public async Task MethodTeardown(CancellationToken cancellationToken)
-    {
-        await Task.Delay(10, cancellationToken);
-        throw new Exception("Method Teardown Exception");
-    }
-}
