--- source/PerformanceTests/ExceptionHandlingExamples/WillThrow.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using Sailfish.Attributes;
-
-namespace PerformanceTests.ExceptionHandlingExamples;
-
-[Sailfish]
-public class WillThrow
-{
-    [SailfishMethod]
-    public void ThisThrows()
-    {
-        throw new Exception("This was my exception!");
-    }
-}
